requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
websocket-client>=1.8.0
aiortc>=1.9.0
av>=12.3.0
numpy>=1.26.0
sounddevice>=0.4.7
soundfile>=0.12.1
coqui-tts>=0.22.0
vosk>=0.3.45
faster-whisper>=1.0.3
tk>=0.1.0
pyautogui>=0.9.54
pywin32>=306; platform_system=="Windows"
openai>=1.33.0
google-generativeai>=0.8.2
webrtcvad>=2.0.10
twilio>=8.10.0
flask>=2.3.0
flask-socketio>=5.3.0
pyngrok>=7.0.0

